Cleaned up detection of physical slots.
Improved error handling. If the kernel does not support the slots directory at
all, we do not complain. Parse errors are now reported as warnings.

Also, the PCI_FILL_PHYS_SLOT is set for all devices whenever we try to detect
the slots, no matter how successful the detection was.