Improve MSI and MSI-X display
There are several ways I think the MSI and MSI-X capabilities could be
reported better in lspci.

Enable is at the beginning of the MSI-X report, but the end of the
MSI report.  Because Enabled is perhaps the most important thing,
I moved it to the start.

The TabSize reported for MSI-X is perhaps a little obscure.  I feel
'Count' works better (and is then the same as MSI).  We can't readily
tell which of the MSI-X capabilities have been programmed, so we can't
report how many have been enabled.

Both capabilities report 'Mask'.  Unfortunately, in MSI, it indicates
whether the device supports per-vector masking, and in MSI-X it indicates
whether all the vectors are masked.  This patch changes both; MSI reports
'Maskable' and MSI-X reports 'Masked'.

All these changes combine to turn the report from this:

        Capabilities: [60] MSI: Mask+ 64bit+ Count=1/8 Enable-
        Capabilities: [78] MSI-X: Enable+ Mask- TabSize=8

to this:

        Capabilities: [60] MSI: Enable- Count=1/8 Maskable+ 64bit+
        Capabilities: [78] MSI-X: Enable+ Count=8 Masked-

which I find much less confusing when juxtaposed.

Signed-off-by: Matthew Wilcox <willy@linux.intel.com>