lspci: Print vendor-/system-specific VPD item ids correctly
Vendor- and system-specific VPD items are matched only by their first
character, so we cannot read the item id from the matching structure.
There is a similar issue with unknown VPD items, which is handled by
creating a matching structure dynamically.

We can handle both cases using the static matching structure and the
actual item id, so do that.

Signed-off-by: Ben Hutchings <bhutchings@solarflare.com>