Improve large BAR display
lspci currently displays anything larger than 2GB as a raw number.
I have a device with a larger BAR and wanted to see that reported as
a multiple of a GB.

I took the opportunity to rewrite this routine to make it easier to add
higher powers in the future.  It's also slightly shorter, which is nice.
I also changed it to use the binary prefixes, much as I dislike them
personally.

Signed-off-by: Matthew Wilcox <willy@linux.intel.com>