lspci: Label PCIe LnkCap "L0s Exit Latency" as "L0s", not "L0"
The PCIe Link Capabilities "L0s Exit Latency" is the latency to exit
L0s, not L0, so label it "L0s" instead of "L0".  This matches the
way we label the Device Capabilities "Endpoint L0s Acceptable Latency"
field as "Latency L0s".  This also adds "Exit" to the description to
help distinguish it from the "Acceptable Latency" fields in the
Device Capabilities register.

Signed-off-by: Bjorn Helgaas <bhelgaas@google.com>