lspci: Clarify "PCIe-to-PCI/PCI-X" desc and Bridge Retry Config Enable
The PCI_EXP_TYPE_PCI_BRIDGE type is a PCIe to PCI/PCI-X bridge, so be a bit
more complete in the comment and printed device type.  Also, per PCIe spec
r3.0, Table 7-14, the PCIe Device Control "Bridge Configuration Retry
Enable" bit only applies to PCIe-to-PCI/PCI-X bridges; it does not apply to
PCI-to-PCIe bridges.

Signed-off-by: Bjorn Helgaas <bhelgaas@google.com>