lspci: Decode PCIe LnkCtl "Read Completion Boundary" for PCIe-to-PCI bridges
The PCIe spec (r3.0, Table 7-16) says the Read Completion Boundary is valid
for Root Ports, Endpoints, and Bridges.  I only added decoding for PCIe-to-
PCI/PCI-X bridges because the RCB of a Bridge indicates the RCB of the
upstream Root Port, so I don't think it makes sense for PCI-to-PCIe
bridges.

Signed-off-by: Bjorn Helgaas <bhelgaas@google.com>