Import initial Darwin Support from Apple
Apple published a patch that adds Darwin support to pciutils. It is not
complete, but it does not appear to cause a build failure on other
platforms. For the sake of attribution purposes, it is being kept
separate from additional changes to make pciutils build properly on
Darwin.

https://www.opensource.apple.com/source/IOPCIFamily/IOPCIFamily-224.92.1/tools/pciutils3.2.0.patch.c

This patch differs slightly from Apple's original patch by omitting the
Makefile.rej that Apple had incorrectly included in the patch that it
placed on its server.

Signed-off-by: Richard Yao <ryao@gentoo.org>