lspci: Avoid printing null character in VPD fields
An 'Unknown' VPD field with 0x0 id is translated to an ascii NULL character
by printf() %c format specifier. This makes for some ugly output and tools
such as gawk will throw a warning when parsing it.

Avoid this by using the %s format specifier with a length limit of 1.

Signed-off-by: Patrick Talbert <ptalbert@redhat.com>