lspci: Add support for the DOE Discovery options
The PCIe 6 specification added support for the Data Object Exchange
(DOE). When DOE is supported the Discovery Data Object Protocol must be
implemented. The protocol allows a requester to obtain information about
the other DOE protocols supported by the device.

This patch adds support for querying the list of DOE Discovery Data
Object Protocols supported by the device.

This unfortunatley does require writing to the device, as we can't
query the information without sending data objects to the device.

Signed-off-by: Alistair Francis <alistair.francis@wdc.com>